APPSERV-36 fixes that alert service no longer was included as data source as it wasn't a server service any more
